create visit

+++ app/src/main/java/org/sic4change/nut4healthcentrotratamiento/data/entitities/Visit.kt
@@ -8,7 +8,7 @@ data class Visit(
     val childId: String,
     val tutorId: String,
     val createdate: Date,
-    val height: Int,
+    val height: Double,
     var weight: Double,
     var imc: Double,
     val armCircunference: Double,
